+++ pkg/mimirtool/commands/rules_test.go
@@ -8,6 +8,8 @@ package commands
 import (
 	"context"
 	"fmt"
+	"os"
+	"path/filepath"
 	"testing"
 
 	"github.com/prometheus/prometheus/model/rulefmt"
@@ -111,6 +113,101 @@ func TestCheckDuplicates(t *testing.T) {
 	} {
 		t.Run(tc.name, func(t *testing.T) {
 			assert.Equal(t, tc.want, checkDuplicates(tc.in))
+		})
+	}
+}
+
+func TestRuleCommand_checkRules(t *testing.T) {
+	completelyBadRuleName := rulefmt.RuleNode{
+		Record: yaml.Node{Value: "up", Kind: yaml.ScalarNode},
+		Expr:   yaml.Node{Value: "up==1", Kind: yaml.ScalarNode},
+	}
+	strictlyBadRuleName := rulefmt.RuleNode{
+		Record: yaml.Node{Value: "up:onecolonmissing", Kind: yaml.ScalarNode},
+		Expr:   yaml.Node{Value: "up==1", Kind: yaml.ScalarNode},
+	}
+	validRule1 := rulefmt.RuleNode{
+		Record: yaml.Node{Value: "rule:up:nothing", Kind: yaml.ScalarNode},
+		Expr:   yaml.Node{Value: "up==1", Kind: yaml.ScalarNode},
+	}
+	validRule2 := rulefmt.RuleNode{
+		Record: yaml.Node{Value: "rule:down:nothing", Kind: yaml.ScalarNode},
+		Expr:   yaml.Node{Value: "up==0", Kind: yaml.ScalarNode},
+	}
+	for _, tc := range []struct {
+		name       string
+		rules      []rulefmt.RuleNode
+		strict     bool
+		shouldFail bool
+	}{
+		{
+			name:       "completely bad rule name, not strict fails too",
+			rules:      []rulefmt.RuleNode{validRule1, completelyBadRuleName, validRule2},
+			strict:     false,
+			shouldFail: true,
+		},
+		{
+			name:       "strictly bad rule name, strict",
+			rules:      []rulefmt.RuleNode{validRule1, strictlyBadRuleName, validRule2},
+			strict:     true,
+			shouldFail: true,
+		},
+		{
+			name:       "strictly bad rule name, not strict",
+			rules:      []rulefmt.RuleNode{validRule1, strictlyBadRuleName, validRule2},
+			strict:     false,
+			shouldFail: false,
+		},
+		{
+			name:       "no duplicates, strict",
+			rules:      []rulefmt.RuleNode{validRule1, validRule2},
+			strict:     true,
+			shouldFail: false,
+		},
+		{
+			name:       "with duplicates, not strict",
+			rules:      []rulefmt.RuleNode{validRule1, validRule1},
+			strict:     false,
+			shouldFail: false,
+		},
+		{
+			name:       "with duplicates, strict",
+			rules:      []rulefmt.RuleNode{validRule1, validRule1},
+			strict:     true,
+			shouldFail: true,
+		},
+	} {
+		t.Run(tc.name, func(t *testing.T) {
+			file := filepath.Join(t.TempDir(), "rules.yaml")
+			{
+				// Setup.
+				f, err := os.Create(file)
+				require.NoError(t, err)
+				t.Cleanup(func() { _ = f.Close() })
+
+				contents := rules.RuleNamespace{
+					Namespace: "test",
+					Groups: []rwrulefmt.RuleGroup{{
+						RuleGroup: rulefmt.RuleGroup{
+							Name:  "rulegroup",
+							Rules: tc.rules,
+						},
+					}},
+				}
+				require.NoError(t, yaml.NewEncoder(f).Encode(contents))
+				require.NoError(t, f.Close())
+			}
+
+			{
+				// Test.
+				cmd := &RuleCommand{Strict: tc.strict, RuleFilesList: []string{file}, Backend: rules.MimirBackend}
+				err := cmd.checkRules(nil)
+				if tc.shouldFail {
+					require.Error(t, err)
+				} else {
+					require.NoError(t, err)
+				}
+			}
 		})
 	}
 }