+++ pkg/storage/tsdb/config.go
@@ -93,7 +93,6 @@ const (
 	headStripeSizeHelp            = "The number of shards of series to use in TSDB (must be a power of 2). Reducing this will decrease memory footprint, but can negatively impact performance."
 	headChunksWriteQueueSizeHelp  = "The size of the write queue used by the head chunks mapper. Lower values reduce memory utilisation at the cost of potentially higher ingest latency. Value of 0 switches chunks mapper to implementation without a queue."
 
-	consistencyDelayFlag                      = "blocks-storage.bucket-store.consistency-delay"
 	maxTSDBOpeningConcurrencyOnStartupFlag    = "blocks-storage.tsdb.max-tsdb-opening-concurrency-on-startup"
 	defaultMaxTSDBOpeningConcurrencyOnStartup = 10
 
@@ -334,19 +333,18 @@ func (cfg *TSDBConfig) IsBlocksShippingEnabled() bool {
 
 // BucketStoreConfig holds the config information for Bucket Stores used by the querier and store-gateway.
 type BucketStoreConfig struct {
-	SyncDir                    string              `yaml:"sync_dir"`
-	SyncInterval               time.Duration       `yaml:"sync_interval" category:"advanced"`
-	MaxConcurrent              int                 `yaml:"max_concurrent" category:"advanced"`
-	TenantSyncConcurrency      int                 `yaml:"tenant_sync_concurrency" category:"advanced"`
-	BlockSyncConcurrency       int                 `yaml:"block_sync_concurrency" category:"advanced"`
-	MetaSyncConcurrency        int                 `yaml:"meta_sync_concurrency" category:"advanced"`
-	DeprecatedConsistencyDelay time.Duration       `yaml:"consistency_delay" category:"deprecated"` // Deprecated. Remove in Mimir 2.9.
-	IndexCache                 IndexCacheConfig    `yaml:"index_cache"`
-	ChunksCache                ChunksCacheConfig   `yaml:"chunks_cache"`
-	MetadataCache              MetadataCacheConfig `yaml:"metadata_cache"`
-	IgnoreDeletionMarksDelay   time.Duration       `yaml:"ignore_deletion_mark_delay" category:"advanced"`
-	BucketIndex                BucketIndexConfig   `yaml:"bucket_index"`
-	IgnoreBlocksWithin         time.Duration       `yaml:"ignore_blocks_within" category:"advanced"`
+	SyncDir                  string              `yaml:"sync_dir"`
+	SyncInterval             time.Duration       `yaml:"sync_interval" category:"advanced"`
+	MaxConcurrent            int                 `yaml:"max_concurrent" category:"advanced"`
+	TenantSyncConcurrency    int                 `yaml:"tenant_sync_concurrency" category:"advanced"`
+	BlockSyncConcurrency     int                 `yaml:"block_sync_concurrency" category:"advanced"`
+	MetaSyncConcurrency      int                 `yaml:"meta_sync_concurrency" category:"advanced"`
+	IndexCache               IndexCacheConfig    `yaml:"index_cache"`
+	ChunksCache              ChunksCacheConfig   `yaml:"chunks_cache"`
+	MetadataCache            MetadataCacheConfig `yaml:"metadata_cache"`
+	IgnoreDeletionMarksDelay time.Duration       `yaml:"ignore_deletion_mark_delay" category:"advanced"`
+	BucketIndex              BucketIndexConfig   `yaml:"bucket_index"`
+	IgnoreBlocksWithin       time.Duration       `yaml:"ignore_blocks_within" category:"advanced"`
 
 	// Chunk pool.
 	DeprecatedMaxChunkPoolBytes           uint64 `yaml:"max_chunk_pool_bytes" category:"deprecated"`             // Deprecated. TODO: Remove in Mimir 2.11.
@@ -410,7 +408,6 @@ func (cfg *BucketStoreConfig) RegisterFlags(f *flag.FlagSet, logger log.Logger) 
 	f.IntVar(&cfg.TenantSyncConcurrency, "blocks-storage.bucket-store.tenant-sync-concurrency", 10, "Maximum number of concurrent tenants synching blocks.")
 	f.IntVar(&cfg.BlockSyncConcurrency, "blocks-storage.bucket-store.block-sync-concurrency", 20, "Maximum number of concurrent blocks synching per tenant.")
 	f.IntVar(&cfg.MetaSyncConcurrency, "blocks-storage.bucket-store.meta-sync-concurrency", 20, "Number of Go routines to use when syncing block meta files from object storage per tenant.")
-	f.DurationVar(&cfg.DeprecatedConsistencyDelay, consistencyDelayFlag, 0, "Minimum age of a block before it's being read. Set it to safe value (e.g 30m) if your object storage is eventually consistent. GCS and S3 are (roughly) strongly consistent.")
 	f.DurationVar(&cfg.IgnoreDeletionMarksDelay, "blocks-storage.bucket-store.ignore-deletion-marks-delay", time.Hour*1, "Duration after which the blocks marked for deletion will be filtered out while fetching blocks. "+
 		"The idea of ignore-deletion-marks-delay is to ignore blocks that are marked for deletion with some delay. This ensures store can still serve blocks that are meant to be deleted but do not have a replacement yet.")
 	f.DurationVar(&cfg.IgnoreBlocksWithin, "blocks-storage.bucket-store.ignore-blocks-within", 10*time.Hour, "Blocks with minimum time within this duration are ignored, and not loaded by store-gateway. Useful when used together with -querier.query-store-after to prevent loading young blocks, because there are usually many of them (depending on number of ingesters) and they are not yet compacted. Negative values or 0 disable the filter.")
@@ -436,9 +433,6 @@ func (cfg *BucketStoreConfig) Validate(logger log.Logger) error {
 	}
 	if err := cfg.MetadataCache.Validate(); err != nil {
 		return errors.Wrap(err, "metadata-cache configuration")
-	}
-	if cfg.DeprecatedConsistencyDelay > 0 {
-		util.WarnDeprecatedConfig(consistencyDelayFlag, logger)
 	}
 	if cfg.DeprecatedMaxChunkPoolBytes != uint64(2*units.Gibibyte) {
 		util.WarnDeprecatedConfig(maxChunksBytesPoolFlag, logger)